+++ enjoy/src/i18n/zh-CN.json
@@ -564,5 +564,6 @@
   "writeNoteHere": "开始做笔记",
   "deleteNote": "删除笔记",
   "areYouSureToDeleteThisNote": "您确定要删除这条笔记吗？",
-  "notesCount": "{{count}} 条笔记"
+  "notesCount": "{{count}} 条笔记",
+  "source": "来源"
 }
