+++ enjoy/src/main.ts
@@ -75,7 +75,7 @@ protocol.registerSchemesAsPrivileged([
 app.on("ready", async () => {
   protocol.handle("enjoy", (request) => {
     let url = request.url.replace("enjoy://", "");
-    if (url.match(/library\/(audios|videos|recordings|speeches)/g)) {
+    if (url.match(/library\/(audios|videos|recordings|speeches|segments)/g)) {
       url = url.replace("library/", "");
       url = path.join(settings.userDataPath(), url);
     } else if (url.startsWith("library")) {
