+++ enjoy/src/renderer/components/posts/post-audio.tsx
@@ -1,12 +1,6 @@
-import { useEffect, useState, useRef, useCallback, useContext } from "react";
+import { useEffect, useState, useContext } from "react";
 import { AppSettingsProviderContext } from "@renderer/context";
-import { renderPitchContour } from "@renderer/lib/utils";
-import { extractFrequencies } from "@/utils";
-import WaveSurfer from "wavesurfer.js";
-import { Button, Skeleton } from "@renderer/components/ui";
-import { PlayIcon, PauseIcon } from "lucide-react";
-import { useIntersectionObserver } from "@uidotdev/usehooks";
-import { secondsToTimestamp } from "@renderer/lib/utils";
+import { Button } from "@renderer/components/ui";
 import { MediaPlayer, MediaProvider } from "@vidstack/react";
 import {
   DefaultAudioLayout,
@@ -16,6 +10,7 @@ import { STORAGE_WORKER_ENDPOINT } from "@/constants";
 import { TimelineEntry } from "echogarden/dist/utilities/Timeline.d.js";
 import { t } from "i18next";
 import { XCircleIcon } from "lucide-react";
+import { WavesurferPlayer } from "../widgets";
 
 export const PostAudio = (props: {
   audio: Partial<MediumType>;
@@ -72,7 +67,8 @@ export const PostAudio = (props: {
         <WavesurferPlayer
           currentTime={currentTime}
           setCurrentTime={setCurrentTime}
-          audio={audio}
+          id={audio.id}
+          src={audio.sourceUrl}
           height={height}
           onError={(err) => setError(err.message)}
         />
@@ -103,138 +99,5 @@ export const PostAudio = (props: {
         </div>
       )}
     </div>
-  );
-};
-
-const WavesurferPlayer = (props: {
-  audio: Partial<MediumType>;
-  height?: number;
-  currentTime: number;
-  setCurrentTime: (currentTime: number) => void;
-  onError?: (error: Error) => void;
-}) => {
-  const { audio, height = 80, onError, setCurrentTime } = props;
-  const [initialized, setInitialized] = useState(false);
-  const [isPlaying, setIsPlaying] = useState(false);
-  const [wavesurfer, setWavesurfer] = useState(null);
-  const containerRef = useRef();
-  const [ref, entry] = useIntersectionObserver({
-    threshold: 1,
-  });
-  const [duration, setDuration] = useState<number>(0);
-
-  const onPlayClick = useCallback(() => {
-    wavesurfer.isPlaying() ? wavesurfer.pause() : wavesurfer.play();
-  }, [wavesurfer]);
-
-  useEffect(() => {
-    // use the intersection observer to only create the wavesurfer instance
-    // when the player is visible
-    if (!entry?.isIntersecting) return;
-    if (!audio.sourceUrl) return;
-    if (wavesurfer) return;
-
-    const ws = WaveSurfer.create({
-      container: containerRef.current,
-      url: audio.sourceUrl,
-      height,
-      barWidth: 1,
-      cursorWidth: 0,
-      autoCenter: true,
-      autoScroll: true,
-      dragToSeek: true,
-      hideScrollbar: true,
-      minPxPerSec: 100,
-      waveColor: "#ddd",
-      progressColor: "rgba(0, 0, 0, 0.25)",
-    });
-
-    setWavesurfer(ws);
-  }, [audio.sourceUrl, entry]);
-
-  useEffect(() => {
-    if (!wavesurfer) return;
-
-    const subscriptions = [
-      wavesurfer.on("play", () => {
-        setIsPlaying(true);
-      }),
-      wavesurfer.on("pause", () => {
-        setIsPlaying(false);
-      }),
-      wavesurfer.on("timeupdate", (time: number) => {
-        setCurrentTime(time);
-      }),
-      wavesurfer.on("ready", () => {
-        setDuration(wavesurfer.getDuration());
-        const peaks = wavesurfer.getDecodedData().getChannelData(0);
-        const sampleRate = wavesurfer.options.sampleRate;
-        const data = extractFrequencies({ peaks, sampleRate });
-        setTimeout(() => {
-          renderPitchContour({
-            wrapper: wavesurfer.getWrapper(),
-            canvasId: `pitch-contour-${audio.id}-canvas`,
-            labels: new Array(data.length).fill(""),
-            datasets: [
-              {
-                data,
-                cubicInterpolationMode: "monotone",
-                pointRadius: 1,
-              },
-            ],
-          });
-        }, 1000);
-        setInitialized(true);
-      }),
-      wavesurfer.on("error", (err: Error) => {
-        onError(err);
-      }),
-    ];
-
-    return () => {
-      subscriptions.forEach((unsub) => unsub());
-      wavesurfer?.destroy();
-    };
-  }, [wavesurfer]);
-
-  return (
-    <>
-      <div className="flex justify-end">
-        <span className="text-xs text-muted-foreground">
-          {secondsToTimestamp(duration)}
-        </span>
-      </div>
-
-      <div
-        ref={ref}
-        className="bg-background rounded-lg grid grid-cols-9 items-center relative h-[80px]"
-      >
-        {!initialized && (
-          <div className="col-span-9 flex flex-col justify-around h-[80px]">
-            <Skeleton className="h-3 w-full rounded-full" />
-            <Skeleton className="h-3 w-full rounded-full" />
-            <Skeleton className="h-3 w-full rounded-full" />
-          </div>
-        )}
-
-        <div className={`flex justify-center ${initialized ? "" : "hidden"}`}>
-          <Button
-            onClick={onPlayClick}
-            className="aspect-square rounded-full p-2 w-12 h-12 bg-blue-600 hover:bg-blue-500"
-          >
-            {isPlaying ? (
-              <PauseIcon className="w-6 h-6 text-white" />
-            ) : (
-              <PlayIcon className="w-6 h-6 text-white" />
-            )}
-          </Button>
-        </div>
-
-        <div
-          className={`col-span-8 ${initialized ? "" : "hidden"}`}
-          ref={containerRef}
-        ></div>
-      </div>
-    </>
   );
 };
