+++ enjoy/src/renderer/components/widgets/index.ts
@@ -1,13 +1,10 @@
 export * from "./db-state";
-
 export * from "./layout";
-export * from "./sidebar";
-export * from "./page-placeholder";
-
-export * from "./login-form";
-
 export * from "./loader-spin";
-export * from "./no-records-found";
-
-export * from "./selection-menu";
 export * from "./lookup-result";
+export * from "./login-form";
+export * from "./no-records-found";
+export * from "./page-placeholder";
+export * from "./selection-menu";
+export * from "./sidebar";
+export * from "./wavesurfer-player";
